+++ src/components/survey/SurveySingleItemView/SurveyComponents/TextViewComponent.tsx
@@ -9,7 +9,7 @@ interface TextViewComponentProps {
   className?: string;
 }
 
-const supportedVariants = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'p', 'li'] as const;
+const supportedVariants = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'p', 'li', 'span'] as const;
 type Variant = typeof supportedVariants[number]; // 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p';
 
 const isValidVariant = (variant: unknown): variant is Variant => {